backend/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.7
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
numpy==1.20.2
pandas==1.2.4
Pillow==8.0.1
promise==2.3
psycopg2-binary==2.8.5
ptvsd==4.3.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.24.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
Unidecode==1.1.1
urllib3==1.25.11
xlrd==1.2.0
